[Refactor] : 테스트용 videousecase 삭제 파일 이동
