+++ app/src/main/java/bootcamp/sparta/disneym/domain/repository/SearchRepositoryImpl.kt
@@ -1,9 +1,7 @@
-package bootcamp.sparta.disneym.repository
+package bootcamp.sparta.disneym.domain.repository
 
 import bootcamp.sparta.disneym.data.datasource.remote.Search
-import bootcamp.sparta.disneym.data.datasource.remote.Videos
 import bootcamp.sparta.disneym.data.datasource.remote.YoutubeRetrofit
-import bootcamp.sparta.disneym.ui.search.SearchRepository
 import retrofit2.Response
 
 /*
@@ -12,11 +10,6 @@ import retrofit2.Response
 * YoutubeRetrofit 객체로 초기화한 API에서 각각의 목록과 정보를 요청하고 반환한다 */
 
 class SearchRepositoryImpl: SearchRepository {
-
-    override suspend fun getVideos(part: String, chart: String, key: String, maxResults: Int, videoCategoryId: Int, regionCode: String) : Response<Videos> {
-        return YoutubeRetrofit.youtubeApi.getPopularVideos(part = part, chart = chart, key = key, maxResults = maxResults, videoCategoryId= videoCategoryId, regionCode = regionCode)
-    }
-
     override suspend fun getSearch(part: String, q: String, maxResults: Int, key: String) : Response<Search>{
         return YoutubeRetrofit.youtubeApi.getSearch(part = part, q = q, maxResults = maxResults, key = key)
     }
