+++ app/src/main/java/bootcamp/sparta/disneym/domain/usecase/mypage/LoadUserDataUseCase.kt
@@ -1,9 +1,9 @@
 package bootcamp.sparta.disneym.domain.usecase.mypage
 
 import android.content.Context
-import bootcamp.sparta.disneym.domain.repository.MyPageRepository
+import bootcamp.sparta.disneym.domain.repository.HomeRepository
 
-class LoadUserDataUseCase (private val repository: MyPageRepository){
+class LoadUserDataUseCase (private val repository: HomeRepository){
 
     operator fun invoke(context: Context) = repository.loadUserData(context)
