+++ app/src/main/java/bootcamp/sparta/disneym/domain/usecase/mypage/UpdateUserIdUseCase.kt
@@ -1,8 +1,8 @@
 package bootcamp.sparta.disneym.domain.usecase.mypage
 
-import bootcamp.sparta.disneym.domain.repository.MyPageRepository
+import bootcamp.sparta.disneym.domain.repository.HomeRepository
 
-class UpdateUserIdUseCase(private val repository: MyPageRepository) {
+class UpdateUserIdUseCase(private val repository: HomeRepository) {
 
     operator fun invoke(id : String) = repository.updateUserId(id)
