+++ app/src/main/java/bootcamp/sparta/disneym/domain/usecase/mypage/UpdateUserProfileUseCase.kt
@@ -1,8 +1,8 @@
 package bootcamp.sparta.disneym.domain.usecase.mypage
 
-import bootcamp.sparta.disneym.domain.repository.MyPageRepository
+import bootcamp.sparta.disneym.domain.repository.HomeRepository
 
-class UpdateUserProfileUseCase(private val repository: MyPageRepository) {
+class UpdateUserProfileUseCase(private val repository: HomeRepository) {
 
     operator fun invoke(profile: Int) = repository.updateUserProfile(profile)
