+++ app/src/main/java/bootcamp/sparta/disneym/domain/usecase/mypage/UpdateUserPwUseCase.kt
@@ -1,8 +1,8 @@
 package bootcamp.sparta.disneym.domain.usecase.mypage
 
-import bootcamp.sparta.disneym.domain.repository.MyPageRepository
+import bootcamp.sparta.disneym.domain.repository.HomeRepository
 
-class UpdateUserPwUseCase(private val repository: MyPageRepository) {
+class UpdateUserPwUseCase(private val repository: HomeRepository) {
 
     operator fun invoke(pw: String) = repository.updateUserPw(pw)
